[#804] Update control centre component
