+++ frontend/src/pages/index.js
@@ -1,6 +1,7 @@
 export { default as Home } from "./home/Home";
 export { default as Forms } from "./forms/Forms";
 export { default as Login } from "./login/Login";
+export { default as ControlCenterLayout } from "./control-center/ControlCenterLayout";
 export { default as ControlCenter } from "./control-center/ControlCenter";
 export { default as Users } from "./users/Users";
 export { default as AddUser } from "./add-user/AddUser";
